Updated Sorting Functions #141
Finished the sort fixes.

- Added ability for moved Wireless_OUT to stay in position.
- Added ability for a tool in between Loader and Wireless_IN to stay in position.
- Removed some redundant sorting calls
- Added some Fus Lib functions.
- While in there, cleaned some type hints.
- Changed some names "Node" to "Tool"

+++ Fusion/Scripts/Libs/Prism_Fusion_lib_Helper.py
@@ -296,7 +296,7 @@ def getFileDataFromAOV(fileList:list, aov:str) -> dict:
 
 #	Returns an average luminance value
 @err_catcher(name=__name__)
-def calculateLuminance(color:dict) -> int:
+def calculateLuminance(color:dict) -> float:
     try:
         r,g,b = color['R'], color['G'], color['B']
         # No need for normalization if RGB values are already in [0, 1]
